adapters/vox-xtts/pyproject.toml
[project]
name = "vox-xtts"
version = "0.2.2"
description = "Coqui XTTS-v2 adapter for Vox"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
    "vox-runtime>=0.2.2",
    "coqui-tts>=0.27.5",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
    "numpy>=1.26.0,<2.4",
    "soundfile>=0.13.1",
]

[project.entry-points."vox.adapters"]
xtts-tts-torch = "vox_xtts.adapter:XTTSAdapter"
xtts = "vox_xtts.adapter:XTTSAdapter"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vox_xtts"]
